LayoutTests/media/video-buffered-too-few-arguments.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description('Test how buffered.start()/end() react to too few arguments.');
var v = document.createElement('video');
shouldThrow("v.buffered.start()");
shouldThrow("v.buffered.end()");
</script>
<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
