Safari-609.1.11.1/LayoutTests/media/video-buffered-too-few-arguments-expected.txt

Test how buffered.start()/end() react to too few arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS v.buffered.start() threw exception TypeError: Not enough arguments.
PASS v.buffered.end() threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
